reduce firmware size by: 1. remove oric-specific buffer and function that was only used in one place (and reusing ReadByte instead) 2. removing redundancy and code duplication in the checkForExt functions 3. removing redundancy and code duplication across the Counter1, Counter2, lcdTime functions 4. replace some ints with bytes 5. move some local 'temporary buffers' outside of functions and reuse (the 'out' buffer for the ReadXxxx functions and the fileheader buffer for tzx and ay) 6. remove some unused global variables
Firmware image size for Arduino Uno build was checked after every change to confirm size reduction.

+++ UEFProcessing.ino
@@ -2,7 +2,7 @@
 void ReadUEFHeader() {
   //Read and check first 12 bytes for a UEF header
   char uefHeader[9];
-  int i=0;
+  byte i=0;
   
   if(entry.seekSet(0)) {
     i = entry.read(uefHeader,9);
